% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server Function for Shiny Application}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Standard Shiny server input parameter}

\item{output}{Standard Shiny server output parameter}

\item{session}{Standard Shiny session parameter}
}
\value{
The function sets up a series of reactive expressions and
  observers to manage app state.
}
\description{
This function contains the server logic for a Shiny application designed to manage and
display data studies. It handles file uploads, data processing, and UI updates based
on the data state.
}
