% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSettingFunction.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply Filters to a Data Frame}
\usage{
apply_filters(df, input, ns, cols_to_filter)
}
\arguments{
\item{df}{A data frame to be filtered.}

\item{input}{Shiny input object containing the filter values.}

\item{ns}{Namespace function for the module.}

\item{cols_to_filter}{A vector of column names to be filtered.}
}
\value{
A filtered data frame.
}
\description{
This function filters a data frame based on the specified columns and input values.
}
