% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{bindData}
\alias{bindData}
\title{Bind a list of data.frames into a single summarised_result object.}
\usage{
bindData(data)
}
\arguments{
\item{data}{A data.frame or list of data.frames.}
}
\value{
A summarised_result object.
}
\description{
Bind a list of data.frames into a single summarised_result object.
}
