% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{cardFromMd}
\alias{cardFromMd}
\title{Create a `bslib::card()` object from a `.md` file.}
\usage{
cardFromMd(fileName)
}
\arguments{
\item{fileName}{Name of the .md file.}
}
\value{
Ui `bslib::card` with the background details.
}
\description{
Create a `bslib::card()` object from a `.md` file.
}
