% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{cardSummary}
\alias{cardSummary}
\title{Create a `bslib::card()` object from a `<summarised_result>` object.}
\usage{
cardSummary(result)
}
\arguments{
\item{result}{A `<summarised_result>` object.}
}
\value{
A `bslib::card()` object.
}
\description{
Create a `bslib::card()` object from a `<summarised_result>` object.
}
