% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSettings.R
\name{correctSettings}
\alias{correctSettings}
\title{Add the strata, group and additional columns to settings with the tidyColumns
(`visOmopResults::splitGroup()`, `visOmopResults::splitStrata()`,
`visOmopResults::splitAdditional()`) for each 'result_id'.}
\usage{
correctSettings(result)
}
\arguments{
\item{result}{A `<summarised_result>` object.}
}
\value{
The summarised_result object with the settings corrected
}
\description{
Add the strata, group and additional columns to settings with the tidyColumns
(`visOmopResults::splitGroup()`, `visOmopResults::splitStrata()`,
`visOmopResults::splitAdditional()`) for each 'result_id'.
}
