% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCard.R
\name{createCard}
\alias{createCard}
\title{Create a Bootstrap Card with Collapsible Content for a Shiny Dashboard}
\usage{
createCard(id, title, setting_filter_ui, filter_ui, table_ui, extra_ui = NULL)
}
\arguments{
\item{id}{The ID for the collapsible card element.}

\item{title}{The title to be displayed on the card header.}

\item{setting_filter_ui}{UI element for the settings filter.}

\item{filter_ui}{UI element for the filter.}

\item{table_ui}{UI element for the table.}

\item{extra_ui}{Optional UI element for additional content, such as a plot (default is NULL).}
}
\value{
A `div` element containing the card structure for use in a Shiny UI.
}
\description{
This function generates a Bootstrap card with a collapsible header and content. The content
includes UI elements for settings, filters, and tables, with an optional extra UI element
that can be displayed in a separate tab.
}
\examples{
\dontrun{
createCard(
  id = "example_card",
  title = "Example Card",
  setting_filter_ui = selectInput("setting", "Setting", choices = c("A", "B")),
  filter_ui = selectInput("filter", "Filter", choices = c("X", "Y")),
  table_ui = DT::dataTableOutput("example_table"),
  extra_ui = plotOutput("example_plot")
)
}
}
