% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPanels.R
\name{defaultPanels}
\alias{defaultPanels}
\title{Default panels defined in the package.}
\usage{
defaultPanels()
}
\value{
Names of the default panels defined in the package.
}
\description{
Default panels defined in the package.
}
\examples{

defaultPanels()

}
