% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{exportStaticApp}
\alias{exportStaticApp}
\title{Export and launch a static shiny specific to the provided results.}
\usage{
exportStaticApp(data = list(), directory = getwd(), launch = TRUE)
}
\arguments{
\item{data}{List of summarised_result objects to build the shiny app.}

\item{directory}{Directory to create the shiny.}

\item{launch}{Whether to launch the shiny app.}
}
\value{
The shiny app will be created in directory.
}
\description{
Export and launch a static shiny specific to the provided results.
}
