% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSettingFunction.R
\name{filter_setting_ui}
\alias{filter_setting_ui}
\title{Table UI Function}
\usage{
filter_setting_ui(id)
}
\arguments{
\item{id}{The module id.}
}
\description{
This function creates the UI for the Shiny module that includes dynamic filters and a data table.
}
