% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{This function is used to filter data in shinys that inputs are defined
following.}
\usage{
filterData(
  result,
  resultType,
  input,
  prefixSet = "",
  prefixGroup = "",
  showSettings = TRUE,
  showGroupping = TRUE,
  pivotEstimates = FALSE
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{resultType}{A resultType of interest to filter by.}

\item{input}{Input of the shiny to filter by.}

\item{prefixSet}{Prefix for setting columns.}

\item{prefixGroup}{Prefix for groupping columns.}

\item{showSettings}{Whether to show settings columns.}

\item{showGroupping}{Whether to show groupping columns.}

\item{pivotEstimates}{Whether to pivot estimates.}
}
\value{
The filtered result.
}
\description{
This function is used to filter data in shinys that inputs are defined
following.
}
