% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSettingFunction.R
\name{filterSettingInitServer}
\alias{filterSettingInitServer}
\title{Initialize Server Logic for Shiny Module}
\usage{
filterSettingInitServer(id, dataset, global_store)
}
\arguments{
\item{id}{The module id.}

\item{dataset}{A reactive expression that returns the dataset to be used.}

\item{global_store}{A reactive value or function to store the unique result IDs.}
}
\description{
This function initializes the server logic for a Shiny module that creates dynamic filters
and displays a filtered data table.
}
