% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSettingFunction.R
\name{filterSettingUi}
\alias{filterSettingUi}
\title{Table UI Function}
\usage{
filterSettingUi(id)
}
\arguments{
\item{id}{The module id.}
}
\description{
This function creates the UI for the Shiny module that includes dynamic filters and a data table.
}
