% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{filterValues}
\alias{filterValues}
\title{Extract filter values.}
\usage{
filterValues(result, panelDetails = NULL)
}
\arguments{
\item{result}{A summarised_result object.}

\item{panelDetails}{A list with the panel details.}
}
\value{
A lit with the values to filter by.
}
\description{
Extract filter values.
}
