% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericUI.R
\name{genericUI}
\alias{genericUI}
\title{Generic UI Function for Dynamic Plot Modules}
\usage{
genericUI(id, moduleType)
}
\arguments{
\item{id}{A string that uniquely identifies the UI instance for namespacing elements.}

\item{moduleType}{A string specifying the type of module, which dictates the UI elements
to generate as defined in `plot_config`.}
}
\value{
A Shiny UI definition that can be integrated into the application layout.
}
\description{
Generic UI Function for Dynamic Plot Modules
}
\details{
The `genericUI` function supports creating a flexible, dynamic user interface in a
Shiny application, aligning with the modular design dictated by `plot_config`. It
automatically generates UI components such as picker inputs for facets, colours, and
other configurable options, as well as output slots for plots and download buttons.
}
