% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{getChoices}
\alias{getChoices}
\title{Get the different options that a summarised_result have.}
\usage{
getChoices(result, flatten = FALSE)
}
\arguments{
\item{result}{A `<summarised_result>` object.}

\item{flatten}{Whether to flatten to a single list or not.}
}
\value{
A named list with the options
}
\description{
Get the different options that a summarised_result have.
}
\examples{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

result <- cdm$cohort1 |>
  summariseCharacteristics()

getChoices(result)

}
