% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPanels.R
\name{getPanel}
\alias{getPanel}
\title{Get one of the default pre-built panels.}
\usage{
getPanel(panelId)
}
\arguments{
\item{panelId}{Name of the one of the default panels. Use \code{defaultPanels()}
to see the available default panels.}
}
\value{
A panel definition.
}
\description{
Get one of the default pre-built panels.
}
\examples{

getPanel("incidence")

}
