% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphCohortOverlap.R
\name{graphCohortOverlap_init_server}
\alias{graphCohortOverlap_init_server}
\title{Server function for the Cohort Overlap Graph Module}
\usage{
graphCohortOverlap_init_server(id, dataset, filter_input)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}

\item{dataset}{A reactive expression that returns the dataset.}

\item{filter_input}{A reactive expression that returns the filter input.}
}
\value{
A module server function.
}
\description{
This function initializes the server-side logic for the cohort overlap graph module.
}
