% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphCohortOverlap.R
\name{graphCohortOverlap_ui}
\alias{graphCohortOverlap_ui}
\title{UI function for the Cohort Overlap Graph Module}
\usage{
graphCohortOverlap_ui(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition for Cohort Overlap plot
}
\description{
This function creates the UI components for the cohort overlap graph module.
}
