% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphCohortTiming.R
\name{graphCohortTiming_ui}
\alias{graphCohortTiming_ui}
\title{UI function for the Cohort Timing Graph Module}
\usage{
graphCohortTiming_ui(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition for Cohort Timing plot
}
\description{
This function creates the UI components for the cohort timing graph module.
}
