% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphIncidence.R
\name{graphIncidence_ui}
\alias{graphIncidence_ui}
\title{UI function for the Incidence Graph Module}
\usage{
graphIncidence_ui(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition for Incidence plot
}
\description{
This function creates the UI components for the Incidence graph module.
}
