% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphLargeScaleCharacteristics.R
\name{graphLargeScaleCharacteristics_ui}
\alias{graphLargeScaleCharacteristics_ui}
\title{UI function for the Large Scale Characteristics Graph Module}
\usage{
graphLargeScaleCharacteristics_ui(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition for large scale characteristics plot
}
\description{
This function creates the UI components for the large scale characteristics graph module.
}
