% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPrevalence.R
\name{graphPrevalence_init_server}
\alias{graphPrevalence_init_server}
\title{Server function for the Prevalence Graph Module}
\usage{
graphPrevalence_init_server(id, dataset, filter_input)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}

\item{dataset}{A reactive expression that returns the dataset.}

\item{filter_input}{A reactive expression that returns the filter input.}
}
\value{
A module server function.
}
\description{
This function initializes the server-side logic for the Prevalence graph module.
}
