% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPrevalence.R
\name{graphPrevalence_ui}
\alias{graphPrevalence_ui}
\title{UI function for the Prevalence Graph Module}
\usage{
graphPrevalence_ui(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition for Prevalence plot
}
\description{
This function creates the UI components for the Prevalence graph module.
}
