% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSummarisedCharacteristics.R
\name{graphSummarisedCharacteristics_ui}
\alias{graphSummarisedCharacteristics_ui}
\title{UI function for the Summarised Characteristics Graph Module}
\usage{
graphSummarisedCharacteristics_ui(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition for summarised characteristics plot
}
\description{
This function creates the UI components for the summarised characteristics graph module.
}
