% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicApp.R
\name{launchDynamicApp}
\alias{launchDynamicApp}
\title{Launch a dynamic shiny app wher you can upload results.}
\usage{
launchDynamicApp()
}
\value{
Launches the shiny app.
}
\description{
Launch a dynamic shiny app wher you can upload results.
}
