% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopViewer-package.R
\docType{package}
\name{omopViewer-package}
\alias{omopViewer}
\alias{omopViewer-package}
\title{omopViewer: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/omopViewer/}
}

}
\author{
\strong{Maintainer}: Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID})

Authors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
