% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopViewerPanels}
\alias{omopViewerPanels}
\title{Panels defined in the package.}
\format{
## `omopViewerPanels`
A list of 26 `omopviewer_panel` obejcts:
}
\usage{
omopViewerPanels
}
\description{
Panels defined in the package.
}
\keyword{datasets}
