% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopViewerResults}
\alias{omopViewerResults}
\title{Mock results obtained from `GiBleed` dataset.}
\format{
## `omopViewerResults`
A `summarised_result` object with 12,017 rows:
}
\usage{
omopViewerResults
}
\description{
Mock results obtained from `GiBleed` dataset.
}
\keyword{datasets}
