% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visOmopResults.R
\name{omopViewerTable}
\alias{omopViewerTable}
\title{Create a formatted gt table.}
\usage{
omopViewerTable(
  result,
  header = character(),
  group = character(),
  hide = character()
)
}
\arguments{
\item{result}{A tibble it must contain at least two columns: estimate_value
and estimate_type.}

\item{header}{Column names to move to header, must be columns in result.}

\item{group}{Column names to group by, must be columns in result.}

\item{hide}{Column names to hide, must be columns in result.}
}
\value{
A gt table.
}
\description{
Create a formatted gt table.
}
