% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPanels.R
\name{panelDetailsFromResult}
\alias{panelDetailsFromResult}
\title{Obtain default panel details from a `<summarised_result>` object.}
\usage{
panelDetailsFromResult(result)
}
\arguments{
\item{result}{A `summarised_result` object.}
}
\description{
Obtain default panel details from a `<summarised_result>` object.
}
