% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare the results for the shiny}
\usage{
prepareData(result)
}
\arguments{
\item{result}{A summarised_result object.}
}
\value{
A list of summarised_result split by result_type.
}
\description{
Split the data by result_type, so it is easier to use inside the shiny app.
}
