% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareShinyData.R
\name{prepareShinyData}
\alias{prepareShinyData}
\title{Prepare a list of summarised_result objects.}
\usage{
prepareShinyData(result, panelDetails = NULL)
}
\arguments{
\item{result}{A summarised_result object}

\item{panelDetails}{A list ith the panel details.}
}
\value{
A list of summarised_result objects.
}
\description{
Prepare a list of summarised_result objects.
}
