% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serverStatic}
\alias{serverStatic}
\title{Provides the static server of the shiny app for a given set of resultType(s).}
\usage{
serverStatic(data = omopgenerics::emptySummarisedResult(), asText = FALSE)
}
\arguments{
\item{asText}{Whether to output a text object or to eval it.}

\item{resultType}{Character vector indicating the result_type of interest.}
}
\value{
The server of interest.
}
\description{
Provides the static server of the shiny app for a given set of resultType(s).
}
