% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarMenuDynamic.R
\name{sidebarMenuDynamic}
\alias{sidebarMenuDynamic}
\title{Generate a Dynamic Sidebar Menu for a Shiny Dashboard}
\usage{
sidebarMenuDynamic(complete_data, incomplet_data)
}
\arguments{
\item{complete_data}{A data frame containing the complete data. It should have columns named `variable_name`, `estimate_name`, and `estimate_value`.}

\item{incomplet_data}{A list or data frame containing the incomplete data.}
}
\value{
A `sidebarMenu` object for a Shiny dashboard.
}
\description{
This function creates a dynamic sidebar menu for a Shiny dashboard application. It includes static menu items and dynamically generated menu items based on the provided complete and incomplete data sets.
}
\examples{
\dontrun{
complete_data <- data.frame(
  variable_name = c("settings", "settings", "other"),
  estimate_name = c("result_type", "result_type", "other"),
  estimate_value = c("table1", "table2", "other_value")
)

incomplet_data <- list(table3 = data.frame(), table4 = data.frame())

sidebar <- sidebarMenuDynamic(complete_data, incomplet_data)
}
}
