% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{splitBySettings}
\alias{splitBySettings}
\title{Split a summarised_result object by a set of setting columns.}
\usage{
splitBySettings(data, columns, collapse = " & ")
}
\arguments{
\item{data}{A summarised_result object.}

\item{columns}{Set of columns to split the summarised_result object by.}

\item{collapse}{Character to collapse names.}
}
\value{
A list of summarised_result objects splitted by `columns`.
}
\description{
Split a summarised_result object by a set of setting columns.
}
