% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySelect.R
\name{studySelectInitServer}
\alias{studySelectInitServer}
\title{Initialize the Study Selection Module Server}
\usage{
studySelectInitServer(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
Initialise the study select module
}
\description{
This function initializes the server-side logic for the study selection module.
}
