% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySelect.R
\name{studySelectServer}
\alias{studySelectServer}
\title{Server function for the Study Selection Module}
\usage{
studySelectServer(input, output, session)
}
\arguments{
\item{input}{Shiny server input.}

\item{output}{Shiny server output.}

\item{session}{Shiny server session.}
}
\value{
A list with a reactive expression that returns the selected study.
}
\description{
This function contains the server-side logic for the study selection module.
}
