% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySelect.R
\name{studySelectUi}
\alias{studySelectUi}
\title{UI function for the Study Selection Module}
\usage{
studySelectUi(id)
}
\arguments{
\item{id}{A string. The namespace identifier for the module.}
}
\value{
A UI definition to select uploaded study
}
\description{
This function creates the UI components for the study selection module.
}
