% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visOmopResults.R
\name{tidyData}
\alias{tidyData}
\title{Get a tidy tibble from a `summarised_result` object.}
\usage{
tidyData(result)
}
\arguments{
\item{result}{A `summarised_result` object.}
}
\value{
A tibble.
}
\description{
Get a tidy tibble from a `summarised_result` object.
}
