% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData.R
\name{uploadData_init_server}
\alias{uploadData_init_server}
\title{Upload Data Server Logic}
\usage{
uploadData_init_server(id, dataFolder)
}
\arguments{
\item{id}{A unique identifier for the Shiny module}

\item{dataFolder}{The folder path to save uploaded data}
}
\description{
This function contains the server-side logic for processing uploads.
}
