% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadData.R
\name{uploadData_ui}
\alias{uploadData_ui}
\title{Upload Data UI Component}
\usage{
uploadData_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny module}
}
\value{
A UI definition for use in a Shiny application.
}
\description{
This function creates a UI component for uploading data files.
}
