% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSettings.R
\name{correctSettings}
\alias{correctSettings}
\title{This function adds infromation on groupping variables to the settings
attribute of a summaried result}
\usage{
correctSettings(x)
}
\arguments{
\item{x}{A `<summarised_result>` object.}
}
\description{
This function adds infromation on groupping variables to the settings
attribute of a summaried result
}
