% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{exportStaticApp}
\alias{exportStaticApp}
\title{Export and launch a static shiny specific to the provided results.}
\usage{
exportStaticApp(
  result = omopgenerics::emptySummarisedResult(),
  directory = getwd(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{directory}{Directory to create the shiny.}

\item{open}{Whether to open the shiny app project.}
}
\value{
The shiny app will be created in directory.
}
\description{
Export and launch a static shiny specific to the provided results.
}
