% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{This function is used to filter data in shinys that inputs are defined
following.}
\usage{
filterData(result, prefix, input)
}
\arguments{
\item{result}{A summarised_result object.}

\item{prefix}{Prefix for the inputs.}

\item{input}{Input of the shiny to filter by.}
}
\value{
The filtered result.
}
\description{
This function is used to filter data in shinys that inputs are defined
following.
}
