% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appDynamic.R
\name{launchDynamicApp}
\alias{launchDynamicApp}
\title{Launch a dynamic shiny app where you can upload results.}
\usage{
launchDynamicApp()
}
\value{
Launches the shiny app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
