% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopViewer-package.R
\docType{package}
\name{omopViewer-package}
\alias{omopViewer}
\alias{omopViewer-package}
\title{omopViewer: Visualise OMOP Results in a Shiny App}
\description{
Visualise results obtained from analysing data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model using Shiny Apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/omopViewer/}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

Authors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID})
}

}
\keyword{internal}
