% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopViewer-package.R
\docType{package}
\name{OmopViewer-package}
\alias{OmopViewer}
\alias{OmopViewer-package}
\title{OmopViewer: Visualise OMOP Results using 'shiny' Applications}
\description{
Visualise results obtained from analysing data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model using 'shiny' applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OmopViewer/}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

Authors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID})
}

Other contributors:
\itemize{
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID}) [contributor]
  \item Nuria Mercade-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID}) [contributor]
  \item Elin Rowlands \email{elin.rowlands@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0005-5166-0417}{ORCID}) [contributor]
  \item Cecilia Campanile \email{cecilia.campanile@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0007-6629-4661}{ORCID}) [contributor]
}

}
\keyword{internal}
