% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopViewerPanels}
\alias{omopViewerPanels}
\title{Panels defined in the package.}
\format{
\subsection{\code{omopViewerPanels}}{

A list of 26 \code{omopviewer_panel} obejcts
}
}
\usage{
omopViewerPanels
}
\description{
The panels are generated using the following file:
\url{https://github.com/OHDSI/OmopViewer/blob/main/data-raw/panels.R}.
}
\keyword{datasets}
