% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopViewerResults}
\alias{omopViewerResults}
\title{Mock results obtained from \code{GiBleed} dataset.}
\format{
\subsection{\code{omopViewerResults}}{

A \code{summarised_result} object with 12,017 rows
}
}
\usage{
omopViewerResults
}
\description{
The results are generated using the following file:
\url{https://github.com/OHDSI/OmopViewer/blob/main/data-raw/results.R}.
}
\keyword{datasets}
