% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopViewerResults}
\alias{omopViewerResults}
\title{Mock results obtained from `GiBleed` dataset.}
\format{
## `omopViewerResults`
A `summarised_result` object with 12,017 rows:
}
\usage{
omopViewerResults
}
\description{
The results are generated using the following file:
<https://github.com/OHDSI/OmopViewer/blob/main/data-raw/results.R>.
}
\keyword{datasets}
