% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{omopViewerThemes}
\alias{omopViewerThemes}
\title{List the available configured \code{OmopViewer} themes}
\usage{
omopViewerThemes()
}
\value{
A character vector with the available \code{OmopViewer} themes.
}
\description{
List the available configured \code{OmopViewer} themes
}
\examples{
library(OmopViewer)

omopViewerThemes()

}
