% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPanels.R
\name{panelDetailsFromResult}
\alias{panelDetailsFromResult}
\title{Obtain default panel details from a \verb{<summarised_result>} object.}
\usage{
panelDetailsFromResult(result)
}
\arguments{
\item{result}{A \code{summarised_result} object.}
}
\value{
A list of \code{omop_viewer_panel} objects.
}
\description{
Obtain default panel details from a \verb{<summarised_result>} object.
}
\examples{

panelDetailsFromResult(omopViewerResults)

}
