% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serverStatic}
\alias{serverStatic}
\title{Provides the static server of the shiny app for a given set of resultType(s).}
\usage{
serverStatic(result = emptySummarisedResult(), asText = FALSE)
}
\arguments{
\item{result}{A summarised_result object.}

\item{asText}{Whether to output a text object or to eval it.}
}
\value{
The server of interest.
}
\description{
Provides the static server of the shiny app for a given set of resultType(s).
}
