% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyData}
\alias{tidyData}
\title{Get a tidy tibble from a `summarised_result` object.}
\usage{
tidyData(result, cols, pivot = "none")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{cols}{Columns to show a part from the core ones.}

\item{pivot}{What to pivot: "none", "estimates" or "estimates and variables".}
}
\value{
A tibble.
}
\description{
Get a tidy tibble from a `summarised_result` object.
}
