% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyData}
\alias{tidyData}
\title{Get a tidy tibble from a `summarised_result` object.}
\usage{
tidyData(
  result,
  showSettings = TRUE,
  showGroupping = TRUE,
  prefixSet = "",
  prefixGroup = "",
  pivotEstimates = FALSE
)
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{showSettings}{Whether to show settings columns.}

\item{showGroupping}{Whether to show groupping columns.}

\item{prefixSet}{Prefix for settings columns.}

\item{prefixGroup}{Prefix for groupping columns.}

\item{pivotEstimates}{Whether to pivot estimates.}
}
\value{
A tibble.
}
\description{
Get a tidy tibble from a `summarised_result` object.
}
