% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyData}
\alias{tidyData}
\title{Get a tidy tibble from a `summarised_result` object.}
\usage{
tidyData(
  result,
  showSettings = TRUE,
  showgrouping = TRUE,
  prefixSet = "",
  prefixGroup = "",
  pivot = "none"
)
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{showSettings}{Whether to show settings columns.}

\item{showgrouping}{Whether to show grouping columns.}

\item{prefixSet}{Prefix for settings columns.}

\item{prefixGroup}{Prefix for grouping columns.}

\item{pivot}{What to pivot: "none", "estimates" or "estimates and variables".}
}
\value{
A tibble.
}
\description{
Get a tidy tibble from a `summarised_result` object.
}
