% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{uiStatic}
\alias{uiStatic}
\title{Provides the static UI of the shiny app for a given set of resultType(s).}
\usage{
uiStatic(result = emptySummarisedResult(), asText = FALSE, logo = NULL)
}
\arguments{
\item{result}{A summarised_result object.}

\item{asText}{Whether to output a text object or to eval it.}

\item{logo}{Whethet to include a logo in the about tab and upper panel.}
}
\value{
The ui of interest.
}
\description{
Provides the static UI of the shiny app for a given set of resultType(s).
}
